refactor: rename `AddPhotoCard` to `add-photo-card`
Renamed related .stories and .module.scss files.
